+++ 360GroupProject/src/ReviewerTest.java
@@ -1,0 +1,36 @@
+
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+import group3project.Reviewer;
+
+public class ReviewerTest {
+
+	Reviewer testReviewer = new Reviewer();
+	
+	public void setup(){
+		
+	}
+	/**
+	 * Test that constructor creates a Reviewer object with fields initialized to zero.
+	 */
+	@Test
+	public void testReviewerConstructor() {
+		assertEquals(testReviewer.getNumReviews(), 0);
+		assertEquals(testReviewer.get)
+	}
+
+	
+	@Test
+	public void testAssignManuscript() {
+		fail("Not yet implemented");
+	}
+	
+	@Test
+	public void testJustAfterDeadline() {
+		
+	}
+
+}
