--- ReviewerTest.java
@@ -1,29 +1,0 @@
-package group3project;
-
-import static org.junit.Assert.*;
-
-import org.junit.Test;
-
-public class ReviewerTest {
-
-	
-	/**
-	 * Test that constructor creates a Reviewer object with fields initialized to zero.
-	 */
-	@Test
-	public void testReviewerConstructor() {
-		fail("Not yet implemented");
-	}
-
-	
-	@Test
-	public void testAssignManuscript() {
-		fail("Not yet implemented");
-	}
-	
-	@Test
-	public void testJustAfterDeadline() {
-		
-	}
-
-}
